Refactor submodules
