+++ fairseq/models/speech_to_speech/s2s_conformer_translatotron2.py
@@ -12,10 +12,12 @@ from fairseq.models import (
     register_model,
     register_model_architecture,
 )
-from fairseq.models.speech_to_speech.modules import CTCDecoder
+from fairseq.models.speech_to_speech.modules.ctc_decoder import CTCDecoder
+from fairseq.models.speech_to_speech.modules.transformer_encoder import (
+    TransformerEncoderNoEmb,
+)
 from fairseq.models.speech_to_speech.s2s_conformer import S2SpecTConformerModel
 from fairseq.models.speech_to_speech.s2s_conformer_unity import (
-    TransformerEncoderNoEmb,
     multitask_text_transformer_decoder_arch,
 )
 from fairseq.models.speech_to_speech.s2s_transformer import (
